Split client from request/response objects
Looking to use the requests package internally if available, else
use default httplib client.
